Add files via upload
Sentiment analysis is part of the Natural Language Processing (NLP) techniques that consists in extracting emotions related to some raw texts. This is usually used on social media posts and customer reviews in order to automatically understand if some users are positive or negative and why. 
In this mini project I will use R for predicting customer reviews if it  is good or bad.